Move Fireplace.java into .device package
