+++ src/main/java/io/patriot_framework/virtual_smart_home/house/HouseBean.java
@@ -1,5 +1,7 @@
 package io.patriot_framework.virtual_smart_home.house;
 
+import io.patriot_framework.virtual_smart_home.house.device.Fireplace;
+
 public class HouseBean {
 
     private House house;
